ImageEdit/app/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bidict==0.23.1
certifi==2025.10.5
click==8.3.0
distlib==0.4.0
distro==1.9.0
docutils==0.22.2
fastapi==0.118.2
filelock==3.19.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
markdown2==2.5.4
MarkupSafe==3.0.3
multidict==6.7.0
nicegui==3.0.3
openai==2.2.0
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.4.0
propcache==0.4.1
pscript==0.7.7
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.1
PyYAML==6.0.3
setuptools==80.9.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.9.0
uvicorn==0.37.0
uvloop==0.21.0
vbuild==0.8.2
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.22.0
